lecture 5: implemented spring data jdbc and removed some boilerplate code

+++ src/main/resources/schema.sql
@@ -3,7 +3,7 @@ create table if not exists fighter (
   name varchar(50) not null,
   damage_per_hit int not null,
   health int not null,
-  resistance numeric not null,
+  resistance double not null,
   anime_from varchar(50) not null,
   created_at varchar(50) not null
 );
